pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dagit>=1.10.12",
    "dagster>=1.10.12",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
]
[tool.dagster]
module_name = "dagster_defs"

